refractoring changes

+++ WorkspaceLauncher/BatchFileCreator.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Text;
+using System.Windows;
 using System.Windows.Controls;
 using WorkspaceLauncher.Models;
 
@@ -10,6 +11,7 @@ namespace WorkspaceLauncher
 {
     internal class BatchFileCreator
     {
+        private static readonly string? ApplicationName = Application.Current.Resources["AppBrandName"] as string;
         public BatchFileCreator()
         {
         }
@@ -37,52 +39,55 @@ namespace WorkspaceLauncher
                 string appDataPath = Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData);
 
                 // Specify the subfolder name
-                string subfolderName = "MyWorkspace"; // Change this to your actual application name
+                string? subfolderName = ApplicationName;
 
                 // Combine the paths to create the full path to the subfolder
-                string subfolderPath = Path.Combine(appDataPath, subfolderName);
-
-                // Check if the subfolder exists, and create it if not
-                if (!Directory.Exists(subfolderPath))
+                if (subfolderName != null)
                 {
-                    Directory.CreateDirectory(subfolderPath);
-                    Debug.WriteLine($"Subfolder '{subfolderName}' created in the Application Data folder.");
-                }
-                else
-                {
-                    Debug.WriteLine($"Subfolder '{subfolderName}' already exists in the Application Data folder.");
-                }
+                    string subfolderPath = Path.Combine(appDataPath, subfolderName);
 
-                //Part 2
-                //Open the output file in write mode
-                Debug.WriteLine("Writing to output file");
-                string outputFilePath = Path.Combine(subfolderPath, outputFileName);
-
-                // Check if the output file exists and delete it
-                if (File.Exists(outputFilePath))
-                {
-                    File.Delete(outputFilePath);
-                    Debug.WriteLine($"Existing file '{outputFilePath}' deleted.");
-                }
-
-                //Part3
-
-                using (StreamWriter outputFile = new StreamWriter(outputFilePath, true, Encoding.UTF8))
-                    //using (StreamWriter outputFile = new StreamWriter(Path.Combine(appDataPath, subfolderName)))
-                {
-                    outputFile.WriteLine("@echo off");
-                    Debug.WriteLine($"Sample file '{outputFilePath}' created in the subfolder.");
-                    foreach (WorkspaceItem? item in listboxItems)
+                    // Check if the subfolder exists, and create it if not
+                    if (!Directory.Exists(subfolderPath))
                     {
-                        string? selectedApplicationPath = item?.ApplicationPath;
-                        string? selectedFilePath = item?.FilePath;
-
-                        // Write the list of files to the text file
-
-                        outputFile.WriteLine(openLinkPrefixCmd + "\"" + selectedApplicationPath + "\" " + "\"" +
-                                             selectedFilePath + "\"" + "\n");
+                        Directory.CreateDirectory(subfolderPath);
+                        Debug.WriteLine($"Subfolder '{subfolderName}' created in the Application Data folder.");
                     }
-                    outputFile.WriteLine("exit");
+                    else
+                    {
+                        Debug.WriteLine($"Subfolder '{subfolderName}' already exists in the Application Data folder.");
+                    }
+
+                    //Part 2
+                    //Open the output file in write mode
+                    Debug.WriteLine("Writing to output file");
+                    string outputFilePath = Path.Combine(subfolderPath, outputFileName);
+
+                    // Check if the output file exists and delete it
+                    if (File.Exists(outputFilePath))
+                    {
+                        File.Delete(outputFilePath);
+                        Debug.WriteLine($"Existing file '{outputFilePath}' deleted.");
+                    }
+
+                    //Part3
+
+                    using (StreamWriter outputFile = new StreamWriter(outputFilePath, true, Encoding.UTF8))
+                        //using (StreamWriter outputFile = new StreamWriter(Path.Combine(appDataPath, subfolderName)))
+                    {
+                        outputFile.WriteLine("@echo off");
+                        Debug.WriteLine($"Sample file '{outputFilePath}' created in the subfolder.");
+                        foreach (WorkspaceItem? item in listboxItems)
+                        {
+                            string? selectedApplicationPath = item?.ApplicationPath;
+                            string? selectedFilePath = item?.FilePath;
+
+                            // Write the list of files to the text file
+
+                            outputFile.WriteLine(openLinkPrefixCmd + "\"" + selectedApplicationPath + "\" " + "\"" +
+                                                 selectedFilePath + "\"" + "\n");
+                        }
+                        outputFile.WriteLine("exit");
+                    }
                 }
 
                 Debug.WriteLine($"File list has been written to {outputFileName}");
